requirements.txt
Django>=4.2,<7.0
mysqlclient>=2.2.0
python-decouple>=3.8
Pillow>=10.0.0
social-auth-app-django>=5.0.0
social-auth-core>=4.4.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
djangorestframework>=3.14.0
PyJWT>=2.8.0
google-generativeai>=0.7.0
openai>=1.0.0
groq>=0.9.0
